+++ bingads/v12/bulk/entities/ad_extensions/common.py
@@ -63,22 +63,22 @@ class _BulkAdExtensionBase(_SingleRecordBulkEntity):
         ),
         _SimpleBulkMapping(
             header=_StringTable.StartDate,
-            field_to_csv=lambda c: field_to_csv_SchedulingStartDate(c._ad_extension.Scheduling),
+            field_to_csv=lambda c: field_to_csv_SchedulingStartDate(c._ad_extension.Scheduling, c._ad_extension.Id),
             csv_to_field=lambda c, v: csv_to_field_Date(c._ad_extension.Scheduling, 'StartDate', v)
         ),
         _SimpleBulkMapping(
             header=_StringTable.EndDate,
-            field_to_csv=lambda c: field_to_csv_SchedulingEndDate(c._ad_extension.Scheduling),
+            field_to_csv=lambda c: field_to_csv_SchedulingEndDate(c._ad_extension.Scheduling, c._ad_extension.Id),
             csv_to_field = lambda c, v: csv_to_field_Date(c._ad_extension.Scheduling, 'EndDate', v)
         ),
         _SimpleBulkMapping(
             header=_StringTable.AdSchedule,
-            field_to_csv=lambda c: field_to_csv_AdSchedule(c._ad_extension.Scheduling),
+            field_to_csv=lambda c: field_to_csv_AdSchedule(c._ad_extension.Scheduling, c._ad_extension.Id),
             csv_to_field=lambda c, v: csv_to_field_AdSchedule(c._ad_extension.Scheduling, v)
         ),
         _SimpleBulkMapping(
             header=_StringTable.UseSearcherTimeZone,
-            field_to_csv=lambda c: field_to_csv_UseSearcherTimeZone(c._ad_extension.Scheduling),
+            field_to_csv=lambda c: field_to_csv_UseSearcherTimeZone(c._ad_extension.Scheduling, c._ad_extension.Id),
             csv_to_field=lambda c, v: setattr(c._ad_extension.Scheduling, 'UseSearcherTimeZone', parse_bool(v))
         ),
         _SimpleBulkMapping(